requirements.txt
Pillow==8.0.1
pytesseract==0.3.6
selenium==3.141.0
opencv-python==4.5.1.48